fix(dsl): сохранить точность арифметики дат
Generated-with: Claude Code

+++ internal/dsl/interpreter/interpreter.go
@@ -1014,7 +1014,7 @@ func dateAddSeconds(t time.Time, sec float64) time.Time {
 	if math.IsNaN(sec) || math.IsInf(sec, 0) || sec > maxWholeSeconds || sec < -maxWholeSeconds {
 		RaiseUserError("сдвиг даты выходит за безопасный диапазон времени")
 	}
-	return t.Add(time.Duration(int64(sec)) * time.Second)
+	return t.Add(time.Duration(sec * float64(time.Second)))
 }
 
 func compare(a, b any) int {
